move all files from pages to componentes and remove page folder from src
